팀원찾기 로직 수정

+++ src/main/java/kr/co/shoulf/web/dto/SearchUserDTO.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class SearchUserDTO {
     private Integer positionNo;
     private Integer positionDetailNo;
-    private Integer onlineNo;
+    private Long onlineNo;
     private String keyword;
 }
